Move type traits to types:: submodule
