+++ crates/chain/chain-components/src/traits/types/proof.rs
@@ -2,7 +2,7 @@ use cgp::prelude::*;
 
 use crate::traits::types::height::HasHeightType;
 
-pub use hermes_chain_type_components::traits::commitment_proof::*;
+pub use hermes_chain_type_components::traits::types::commitment_proof::*;
 
 #[derive_component(CommitmentProofHeightGetterComponent, CommitmentProofHeightGetter<Chain>)]
 pub trait HasCommitmentProofHeight: HasCommitmentProofType + HasHeightType {
