+++ crates/chain/chain-type-components/src/traits/mod.rs
@@ -1,8 +1,1 @@
-pub mod commitment_prefix;
-pub mod commitment_proof;
-pub mod event;
-pub mod height;
-pub mod ibc;
-pub mod message;
-pub mod time;
-pub mod timeout;
+pub mod types;
