+++ crates/chain/chain-type-components/src/traits/types/commitment_proof.rs
