+++ crates/chain/chain-type-components/src/traits/types/mod.rs
@@ -1,0 +1,8 @@
+pub mod commitment_prefix;
+pub mod commitment_proof;
+pub mod event;
+pub mod height;
+pub mod ibc;
+pub mod message;
+pub mod time;
+pub mod timeout;
